fix: move topics to sdk folder
